moving/requirements.txt
pandas==1.5.3
scikit-learn==1.2.2
numpy==1.23.5
matplotlib==3.7.0
reformer-pytorch==1.4.4
datetime==5.5